Add ChangeColorBar fade effect
- FadeEffcet: SpriteRenderer 알파를 보간하는 재사용 가능한 코루틴 유틸리티 추가
- ChangeColorBarController: 자식 SpriteRenderer들을 반복 페이드 인/아웃시키는 컨트롤러 추가
- ChangeColorBar 프리팹 추가
- CameraOrthographicSize.cs 포맷 정리

+++ My project (1)/Assets/Scripts/CameraOrthographicSize.cs
@@ -2,14 +2,15 @@ using UnityEngine;
 
 public class CameraOrthographicSize : MonoBehaviour
 {
-private void Awake()
+    private void Awake()
     {
         float screenAspectRatio = (float)Screen.width / Screen.height;
-        float orthographicSize = (6-(screenAspectRatio-0.485f) * 11f);
-        if(orthographicSize < 4)
+        float orthographicSize = (float)(6 - (screenAspectRatio - 0.485f) * 11f);
+        if (orthographicSize < 4)
         {
             orthographicSize = 4;
         }
+
         Camera.main.orthographicSize = orthographicSize;
     }
 }
